Split table output tests from renderer
